backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2
email-validator==2.2.0
python-multipart==0.0.17
pytest==8.3.4
httpx==0.28.1
